refactoring, tests for service deployment
